Use unliftio instead of enclosed-exceptions
`UnliftIO.Exception` has the useful `tryAny`/`catchAny` functions.

It also has `forkIO`, `bracket`, `modifyMVar`, etc. which work for
`ReaderT r IO`, which makes things much easier.

So enclosed-exceptions is replaced by unliftio -- a package with a
much broader remit.

Unliftio is convenient and mostly harmless. The only thing to watch
out for is that its versions of `bracket` and `finally` use
`uninterruptibleMask` in the cleanup functions.

Where UnliftIO modules are imported, I have added import lists, so
that it's clear where symbols come from.

Qualified imports of MVar and Exception are changed to unqualified
imports because there is no ambiguity about the symbols imported.

Some uses of `liftIO` become redundant when used with unliftio
functions, so I have changed those places.

The `taffyFork` function is now mostly unnecessary because the
`MonadUnliftIO` version of `forkIO` can be used with any ReaderT.